removed page folder and refactored accordingly
