+++ frontend/src/App.tsx
@@ -1,20 +1,23 @@
 import { Navigate, Route, Routes } from 'react-router-dom';
 
-import { AttestationPage } from './features/attestation/AttestationPage';
-import { AppLayout, Auth, Chat, Overview, User } from './pages';
+import { AttestationPage } from './features/attestation';
+import { AuthPage } from './features/auth';
+import { ChatPage } from './features/chat';
+import { UserPage } from './features/user';
+import { AppLayout, OverviewPage } from './ui';
 
 export function App() {
     return (
         <div className="font-quicksand flex h-full w-full">
             <Routes>
                 <Route path="/" element={<Navigate to="/login" replace />} />
-                <Route path="/login" element={<Auth />} />
+                <Route path="/login" element={<AuthPage />} />
 
                 {/* TODO: Add <ProtectedRoute/> */}
                 <Route path="/app" element={<AppLayout />}>
-                    <Route index element={<Overview />} />
-                    <Route path="chat" element={<Chat />} />
-                    <Route path="user" element={<User />} />
+                    <Route index element={<OverviewPage />} />
+                    <Route path="chat" element={<ChatPage />} />
+                    <Route path="user" element={<UserPage />} />
                     <Route path="attestation" element={<AttestationPage />} />
                 </Route>
             </Routes>
