+++ frontend/src/features/auth/index.ts
@@ -2,3 +2,4 @@ export * from './AuthForm';
 export * from './AuthInput';
 export * from './AuthLayout';
 export * from './AuthToggle';
+export * from './AuthPage';
