+++ frontend/src/features/chat/ChatPage.tsx
@@ -1,6 +1,6 @@
-import { ChatHeader, ChatInput, ChatMessages } from '../features/chat';
+import { ChatHeader, ChatInput, ChatMessages } from '.';
 
-export function Chat() {
+export function ChatPage() {
     return (
         <main className="flex flex-1 flex-col bg-gradient-to-r from-teal-950 from-2% to-slate-800 to-95%">
             <ChatHeader />
