+++ frontend/src/features/chat/index.ts
@@ -1,3 +1,4 @@
 export * from './ChatHeader';
 export * from './ChatMessages';
 export * from './ChatInput';
+export * from './ChatPage';
