+++ frontend/src/features/user/UserPage.tsx
@@ -1,4 +1,4 @@
-export function User() {
+export function UserPage() {
     return (
         <main className="flex flex-1 flex-col items-center justify-center space-y-2 bg-gradient-to-r from-teal-950 from-2% to-slate-800 to-95%">
             {/* TODO: Add load, delete models for the user session. */}
