+++ frontend/src/features/user/index.ts
@@ -1,1 +1,2 @@
 export * from './UserSummary';
+export * from './UserPage';
