--- frontend/src/pages/index.ts
@@ -1,5 +1,0 @@
-export * from './AppLayout';
-export * from './Auth';
-export * from './Chat';
-export * from './User';
-export * from './Overview';
