+++ frontend/src/ui/index.ts
@@ -1,3 +1,5 @@
 export * from './Logo';
 export * from './ProfilePicture';
 export * from './Modal';
+export * from './AppLayout';
+export * from './OverviewPage';
